requirements-ci.txt
#lint
pycodestyle==2.5.0

# Test
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0